+++ packages/typescript/src/index.js
@@ -1,0 +1,6 @@
+//--------------------------------------------------------
+//-- ESLint config
+//--------------------------------------------------------
+import loader from "@absolunet/eslint-loader";
+
+module.exports = loader(__dirname);