+++ packages/typescript/test/feature/index.test.js
@@ -1,0 +1,9 @@
+//--------------------------------------------------------
+//-- typescript - Feature tests
+//--------------------------------------------------------
+import configTests from "../../../../test/feature/config-tests";
+import { given, when, then } from "./index.gwt";
+
+describe(`Validate that 'typescript' works`, () => {
+	configTests({ given, when, then });
+});