+++ packages/typescript/test/generic/index.test.js
@@ -1,0 +1,6 @@
+//--------------------------------------------------------
+//-- Tester
+//--------------------------------------------------------
+import { tester } from "@absolunet/tester";
+
+tester.genericRepositoryTests();