+++ test/feature/config.gwt.js
@@ -2,6 +2,7 @@
 //-- Config - Given-When-Then
 //--------------------------------------------------------
 import path from "path";
+import deepmerge from "deepmerge";
 import { given as baseGiven, when as baseWhen, then as baseThen } from "../base.gwt.js";
 
 const given = { ...baseGiven };
@@ -14,6 +15,7 @@ const { ESLint } = require(require.resolve("eslint", {
 
 let root;
 let config;
+let extraConfig;
 let eslint;
 
 //-- Given - Reset
@@ -25,6 +27,10 @@ given.noConfig = () => {
 	config = undefined;
 };
 
+given.noExtraConfig = () => {
+	extraConfig = undefined;
+};
+
 given.noESLint = () => {
 	eslint = undefined;
 };
@@ -34,6 +40,16 @@ given.root = (value) => {
 	root = value;
 };
 
+//-- Given - Extra config
+given.extraConfig = (value) => {
+	extraConfig = value;
+};
+
+given.currentExtraConfig = () => {
+	// No extra config by default
+	extraConfig = {};
+};
+
 //-- When - Config
 when.configIsLoaded = () => {
 	when.attempting(() => {
@@ -41,8 +57,13 @@ when.configIsLoaded = () => {
 	});
 };
 
+when.extraConfigIsMerged = () => {
+	config = deepmerge(config, extraConfig);
+};
+
 when.configIsParsed = () => {
 	when.configIsLoaded();
+	when.extraConfigIsMerged();
 
 	when.attempting(() => {
 		eslint = new ESLint({